refactor(Card, Dropdown): 코드 리팩토링
