+++ src/components/card/CardCreateModal.tsx
@@ -1,15 +1,15 @@
 /***이서님이 수정해주실 페이지 !****/
 
-import { useEffect, useRef, useState } from 'react';
+import { ChangeEvent, useEffect, useRef, useState } from 'react';
 import UserDropdown from '../dropdown/UserDropdown';
 import { ModalRoot } from '../modal/ModalRoot';
 import { getMembersApi, members, getUserMeAPI, PostCard, CardRequest } from '@/api/card/apis';
-import { UserType } from '../chip/UserChip';
 import { Input } from '../input/Input';
 import DatePicker from 'react-datepicker';
-import { useImageUpload, imageUrl } from '@/hooks/useImageUpload';
+import { useImageUpload } from '@/hooks/useImageUpload';
 import ImageUpload from '../ImageUpload';
 import { formatDueDate } from '@/utils/formatDueDate';
+import { UserType } from '@/types/UserTypes';
 
 interface Props {
   modalOpenSetState: (state: boolean) => void;
@@ -74,12 +74,7 @@ const CardCreateModal = ({
         const res = await getMembersApi(dashboardId);
         const data = res.members;
         const res2 = await getUserMeAPI();
-        const data2 = {
-          id: res2.id,
-          nickname: res2.nickname,
-          profileImageUrl: res2.profileImageUrl,
-        };
-        setUserData(data2);
+        setUserData(res2);
         setMembers(data);
       } catch (err) {
         console.error('멤버 가져오기 실패:', err);
@@ -87,7 +82,7 @@ const CardCreateModal = ({
     };
     handleGetMembers();
   }, []);
-  const titleSetting = (e) => {
+  const titleSetting = (e: ChangeEvent<HTMLInputElement>) => {
     setTitle(e.target.value);
   };
   const handleDueDateChange = (date: Date | null) => {
@@ -141,14 +136,9 @@ const CardCreateModal = ({
         >
           <UserDropdown.Trigger>이름을 입력해 주세요</UserDropdown.Trigger>
           <UserDropdown.Content>
-            {members.map((user) => {
-              const converted: UserType = {
-                id: user.id,
-                nickname: user.nickname,
-                profileImageUrl: user.profileImageUrl,
-              };
-              return <UserDropdown.Item key={user.id}>{converted}</UserDropdown.Item>;
-            })}
+            {members.map((user) => (
+              <UserDropdown.Item key={user.id}>{user}</UserDropdown.Item>
+            ))}
           </UserDropdown.Content>
         </UserDropdown.Root>
         <label className='block mb-2 text-sm font-medium text-gray-700'>
